requirements.txt
# Core pipeline
datasets>=2.20.0
transformers>=4.44.0
accelerate>=0.33.0
sentence-transformers>=3.0.0
pyyaml>=6.0.1
pydantic>=2.7.0
tqdm>=4.66.0
numpy>=1.26.0
scikit-learn>=1.4.0
rapidfuzz>=3.9.0
python-dotenv>=1.0.1

# Bundled evaluation and utility scripts
google-genai>=1.65.0
requests>=2.32.0
lxml>=5.0.0

# Optional acceleration on Linux
bitsandbytes>=0.43.0; platform_system != "Windows"
